requirements.txt
langchain>=0.3.0
langchain-community
langchain-openai
langchain-chroma
langchain-experimental
pypdf
streamlit
python-dotenv
duckduckgo-search
langchain-google-genai
fastapi
uvicorn
python-multipart
requests
Pillow
pytesseract
# Optional for full image OCR: pip install "unstructured[all-docs]"
